+++ src/main/java/uk/q3c/krail/core/option/VaadinOptionContext.java
@@ -1,0 +1,9 @@
+package uk.q3c.krail.core.option;
+
+import com.vaadin.data.Property;
+
+/**
+ * Created by David Sowerby on 03 Aug 2017
+ */
+public interface VaadinOptionContext extends OptionContext<Property.ValueChangeEvent> {
+}