+++ src/main/java/uk/q3c/krail/core/persist/cache/i18n/DefaultPatternCacheLoader.java
@@ -19,8 +19,8 @@ import com.vaadin.data.Property;
 import uk.q3c.krail.core.i18n.DescriptionKey;
 import uk.q3c.krail.core.i18n.LabelKey;
 import uk.q3c.krail.core.option.Option;
-import uk.q3c.krail.core.option.OptionContext;
 import uk.q3c.krail.core.option.OptionKey;
+import uk.q3c.krail.core.option.VaadinOptionContext;
 import uk.q3c.krail.i18n.DefaultPatternSourceProvider;
 import uk.q3c.krail.i18n.I18NModule;
 import uk.q3c.krail.i18n.KrailResourceBundleControl;
@@ -43,7 +43,7 @@ import static com.google.common.base.Preconditions.*;
  * PatternSourceProvider}.  David Sowerby 26/07/15
  * Created by David Sowerby on 08/12/14.
  */
-public class DefaultPatternCacheLoader extends CacheLoader<PatternCacheKey, String> implements PatternCacheLoader, OptionContext {
+public class DefaultPatternCacheLoader extends CacheLoader<PatternCacheKey, String> implements PatternCacheLoader, VaadinOptionContext {
     public static final OptionKey<Boolean> optionKeyAutoStub = new OptionKey<>(Boolean.FALSE, DefaultPatternSourceProvider.class, LabelKey.Auto_Stub,
             DescriptionKey.Auto_Stub);
     public static final OptionKey<Boolean> optionKeyStubWithKeyName = new OptionKey<>(Boolean.TRUE, DefaultPatternSourceProvider.class, LabelKey