+++ src/main/java/uk/q3c/krail/i18n/DefaultPatternSourceProvider.java
@@ -17,7 +17,6 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-import com.vaadin.data.Property;
 import uk.q3c.krail.core.option.AnnotationOptionList;
 import uk.q3c.krail.core.option.Option;
 import uk.q3c.krail.i18n.api.I18NKey;
@@ -85,7 +84,7 @@ public class DefaultPatternSourceProvider implements PatternSourceProvider {
      * {@inheritDoc}
      */
     @Override
-    public void optionValueChanged(Property.ValueChangeEvent event) {
+    public void optionValueChanged(Object event) {
         //do nothing options used as required
     }