+++ src/test/java/uk/q3c/krail/core/option/MockContext0.java
@@ -1,6 +1,5 @@
 package uk.q3c.krail.core.option;
 
-import com.vaadin.data.Property;
 import uk.q3c.krail.testutil.option.MockOption;
 
 import javax.annotation.Nonnull;
@@ -19,7 +18,7 @@ public class MockContext0 extends MockContext {
     }
 
     @Override
-    public void optionValueChanged(Property.ValueChangeEvent event) {
+    public void optionValueChanged(Object event) {
 
     }
 }